+++ sdk/ai/azure-ai-agents/tests/test_agents_client_async.py
@@ -22,6 +22,7 @@ from azure.ai.agents.models import (
     AzureFunctionStorageQueue,
     AgentStreamEvent,
     AgentThread,
+    BingCustomSearchTool,
     BingGroundingTool,
     BrowserAutomationTool,
     CodeInterpreterTool,
@@ -46,6 +47,8 @@ from azure.ai.agents.models import (
     ResponseFormatJsonSchemaType,
     RunAdditionalFieldList,
     RunStepDeltaAzureAISearchToolCall,
+    RunStepDeltaCustomBingGroundingToolCall,
+    RunStepBingCustomSearchToolCall,
     RunStepBingGroundingToolCall,
     RunStepBrowserAutomationToolCall,
     RunStepConnectedAgentToolCall,
@@ -3220,12 +3223,12 @@ class TestAgentClientAsync(TestAgentClientBase):
         """Test Bing grounding tool call in non-streaming Scenario."""
         async with self.create_client(by_endpoint=True, **kwargs) as client:
             model_name = "gpt-4o"
-            openapi_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
+            bing_grounding_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
 
             await self._do_test_tool(
                 client=client,
                 model_name=model_name,
-                tool_to_test=openapi_tool,
+                tool_to_test=bing_grounding_tool,
                 instructions="You are helpful agent",
                 prompt="How does wikipedia explain Euler's Identity?",
                 expected_class=RunStepBingGroundingToolCall,
@@ -3241,15 +3244,63 @@ class TestAgentClientAsync(TestAgentClientBase):
         """Test Bing grounding tool call in streaming Scenario."""
         async with self.create_client(by_endpoint=True, **kwargs) as client:
             model_name = "gpt-4o"
-            openapi_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
+            bing_grounding_tool = BingGroundingTool(connection_id=kwargs.get("azure_ai_agents_tests_bing_connection_id"))
 
             await self._do_test_tool_streaming(
                 client=client,
                 model_name=model_name,
-                tool_to_test=openapi_tool,
+                tool_to_test=bing_grounding_tool,
                 instructions="You are helpful agent",
                 prompt="How does wikipedia explain Euler's Identity?",
                 expected_delta_class=RunStepDeltaBingGroundingToolCall,
+                uri_annotation=MessageTextUrlCitationDetails(
+                    url="*",
+                    title="*",
+                ),
+            )
+    
+    @agentClientPreparer()
+    @recorded_by_proxy_async
+    async def test_custom_bing_grounding_tool(self, **kwargs):
+        """Test Bing grounding tool call in non-streaming Scenario."""
+        async with self.create_client(by_endpoint=True, **kwargs) as client:
+            model_name = "gpt-4o"
+            bing_custom_tool = BingCustomSearchTool(
+                connection_id=kwargs.get("azure_ai_agents_tests_bing_custom_connection_id"),
+                instance_name=kwargs.get("azure_ai_agents_tests_bing_configuration_name")
+            )
+
+            await self._do_test_tool(
+                client=client,
+                model_name=model_name,
+                tool_to_test=bing_custom_tool,
+                instructions="You are helpful agent",
+                prompt="How many medals did the USA win in the 2024 summer olympics?",
+                expected_class=RunStepBingCustomSearchToolCall,
+                uri_annotation=MessageTextUrlCitationDetails(
+                    url="*",
+                    title="*",
+                ),
+            )
+
+    @agentClientPreparer()
+    @recorded_by_proxy_async
+    async def test_custom_bing_grounding_tool_streaming(self, **kwargs):
+        """Test Bing grounding tool call in streaming Scenario."""
+        async with self.create_client(by_endpoint=True, **kwargs) as client:
+            model_name = "gpt-4o"
+            bing_custom_tool = BingCustomSearchTool(
+                connection_id=kwargs.get("azure_ai_agents_tests_bing_custom_connection_id"),
+                instance_name=kwargs.get("azure_ai_agents_tests_bing_configuration_name")
+            )
+
+            await self._do_test_tool_streaming(
+                client=client,
+                model_name=model_name,
+                tool_to_test=bing_custom_tool,
+                instructions="You are helpful agent",
+                prompt="How many medals did the USA win in the 2024 summer olympics?",
+                expected_delta_class=RunStepDeltaCustomBingGroundingToolCall,
                 uri_annotation=MessageTextUrlCitationDetails(
                     url="*",
                     title="*",